Moved StringExtensions methods: CommonPlayniteShared -> CommonPluginsShared
