+++ CommonPlayniteShared/Common/Extensions/StringExtensions.cs
@@ -14,43 +14,6 @@ namespace System
     {
         private static readonly CultureInfo enUSCultInfo = new CultureInfo("en-US", false);
 
-        private static Regex TrimmableWhitespace = new Regex(@"^\s+|\s+$", RegexOptions.Compiled);
-        private static Regex NonWordCharactersAndTrimmableWhitespace = new Regex(@"(?<start>^[\W_]+)|(?<end>[\W_]+$)|(?<middle>[\W_]+)", RegexOptions.Compiled);
-
-        public static string TrimWhitespace(this string input)
-        {
-            return TrimmableWhitespace.Replace(input, "");
-        }
-
-        public static string NormalizeTitleForComparison(this string title)
-        {
-            MatchEvaluator matchEvaluator = (Match match) =>
-            {
-                if (match.Groups["middle"].Success) //if the match group is the last one in the regex (non-word characters, including whitespace, in the middle of a string)
-                    return " ";
-                else
-                    return ""; //remove non-word characters (including white space) at the start and end of the string
-            };
-            return NonWordCharactersAndTrimmableWhitespace.Replace(title, matchEvaluator).RemoveDiacritics();
-        }
-
-        public static string RemoveDiacritics(this string text)
-        {
-            var normalizedString = text.Normalize(NormalizationForm.FormD);
-            var stringBuilder = new StringBuilder();
-
-            foreach (var c in normalizedString)
-            {
-                var unicodeCategory = CharUnicodeInfo.GetUnicodeCategory(c);
-                if (unicodeCategory != UnicodeCategory.NonSpacingMark)
-                {
-                    stringBuilder.Append(c);
-                }
-            }
-
-            return stringBuilder.ToString().Normalize(NormalizationForm.FormC);
-        }
-
         public static string MD5(this string s)
         {
             using (var provider = System.Security.Cryptography.MD5.Create())
